Fix MC/ESME deadlock in TRX mode.
An attemt to fix closed but not resolved issue #1. When sending pdus
from gen_mc/gen_esme to sessions, events to session fsms are sent in a
syncronous way. This causes deadlocks when there is an incoming pdu,
which is simultaneously and also syncronously passed from a session fsm
to gen_mc/gen_esme.

But actually sending pdus to sessions doesn't need to be syncronous: the
only result from session fsm is a unique ref, which is used further for
passing to callbacks. This ref can be generated and returned
immediately, and thus the entire outgoing pdu with the attached ref can
be sent to the session fsm asyncronously. An attempt to make such a fix
is contained in this PR. Мy generic examples which got constantly
blocked (on both mc and esme sides) stopped getting blocked after fix.

Author:     av <av@fun-box.ru>
Repository: savonarola
Commit:	    8ed55c33e49397497aea0ae56ae94d473ccc9e02